pyproject.toml
[tool.poetry]
name = "dust-modeling"
version = "0.1.0"
description = ""
authors = ["kkin1995"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.3"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
astropy = "^6.0.0"
python-dotenv = "^1.0.0"
scipy = "^1.11.4"
pyyaml = "^6.0.1"
dustmaps = "^1.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
